Add jackson as alternative serializer. Needs polishing e.g. remove the RuntimeExceptions.
