+++ reactive-spring-session-postgres/src/main/java/net/andreaskluth/session/postgres/serializer/JacksonSerializationStrategy.java
@@ -1,0 +1,5 @@
+package net.andreaskluth.session.postgres.serializer;
+
+public class JacksonSerializationStrategy {
+
+}
