+++ reactive-spring-session-postgres/src/main/java/net/andreaskluth/session/postgres/serializer/SerializationStrategy.java
@@ -1,0 +1,5 @@
+package net.andreaskluth.session.postgres.serializer;
+
+public interface SerializationStrategy {
+
+}
