+++ reactive-spring-session-postgres/src/test/java/net/andreaskluth/session/postgres/serializer/JacksonSerializationStrategyTest.java
@@ -1,0 +1,7 @@
+package net.andreaskluth.session.postgres.serializer;
+
+import static org.junit.Assert.*;
+
+public class JacksonSerializationStrategyTest {
+
+}
