+++ reactive-spring-session-postgres/src/test/java/net/andreaskluth/session/postgres/serializer/TestSerializationStrategyBase.java
@@ -1,0 +1,5 @@
+package net.andreaskluth.session.postgres.serializer;
+
+public class TestSerializationStrategyBase {
+
+}
